Split msh example to new file
